split core Client trait into Invoker + WrapLoader + WrapInvoker + UriResolverHandler
