+++ packages/client/src/subinvoker.rs
@@ -40,4 +40,7 @@ impl Invoker for Subinvoker {
     fn get_interfaces(&self) -> Option<InterfaceImplementations> {
         self.invoker.get_interfaces()
     }
+    fn get_env_by_uri(&self, uri: &Uri) -> Option<&Env> {
+        self.invoker.get_env_by_uri(uri)
+    }
 }
