+++ packages/core/src/client.rs
@@ -1,13 +1,12 @@
 use std::sync::Arc;
 
-use crate::error::Error;
-use crate::invoker::Invoker;
-use crate::resolvers::uri_resolution_context::UriResolutionContext;
 use crate::uri::Uri;
 use crate::interface_implementation::InterfaceImplementations;
 use crate::resolvers::uri_resolver::{UriResolverHandler, UriResolver};
-use crate::env::{Envs, Env};
-use crate::wrapper::Wrapper;
+use crate::env::Envs;
+use crate::invoker::Invoker;
+use crate::wrap_invoker::WrapInvoker;
+use crate::wrap_loader::WrapLoader;
 
 #[derive(Clone,Debug)]
 pub struct UriRedirect {
@@ -28,20 +27,4 @@ pub struct ClientConfig {
   pub interfaces: Option<InterfaceImplementations>
 }
 
-pub trait Client: Invoker + UriResolverHandler {
-  fn get_env_by_uri(&self, uri: &Uri) -> Option<&Env>;
-  fn load_wrapper(
-    &self,
-    uri: &Uri,
-    resolution_context: Option<&mut UriResolutionContext>,
-  ) -> Result<Arc<dyn Wrapper>, Error>;
-  fn invoke_wrapper_raw(
-    &self,
-    wrapper: &dyn Wrapper,
-    uri: &Uri,
-    method: &str,
-    args: Option<&[u8]>,
-    env: Option<&Env>,
-    resolution_context: Option<&mut UriResolutionContext>,
-) -> Result<Vec<u8>, Error>;
-}
+pub trait Client: Invoker + WrapLoader + WrapInvoker + UriResolverHandler {}
