+++ packages/core/src/invoker.rs
@@ -1,7 +1,7 @@
 use crate::{
-    error::Error, uri::Uri, resolvers::uri_resolution_context::UriResolutionContext, env::{Env}, interface_implementation::InterfaceImplementations,
+    error::Error, uri::Uri, resolvers::uri_resolution_context::UriResolutionContext, env::Env, 
+    interface_implementation::InterfaceImplementations,
 };
-
 pub trait Invoker: Send + Sync {
     fn invoke_raw(
         &self,
@@ -13,4 +13,5 @@ pub trait Invoker: Send + Sync {
     ) -> Result<Vec<u8>, Error>;
     fn get_implementations(&self, uri: &Uri) -> Result<Vec<Uri>, Error>;
     fn get_interfaces(&self) -> Option<InterfaceImplementations>;
+    fn get_env_by_uri(&self, uri: &Uri) -> Option<&Env>;
 }
